python-app/requirements.txt
# https://pypi.org/project/questionary/
questionary==2.0.1
# https://pypi.org/project/PyYAML/
PyYAML==6.0.2
# https://pypi.org/project/Jinja2/
jinja2==3.1.4
# https://pypi.org/project/termcolor/
termcolor==2.5.0
#https://pypi.org/project/Cerberus/
Cerberus==1.3.5
